[ganetimgr1.8] Fix issues with migrations, add new ones
Actions:

* Delete old south migrations
* Add new (initial) migrations for apps
* Refactor `ApplicationAdmin` to `InstanceApplicationAdmin`
* Remove dynamic `ListView` actions for `InstanceApplicationAdmin`
* Refactor `ManyToManyField`s, remove `null=True` (no effect, triggers
    warning in Django>=1.8), add required migrations

Note: This is a combination of different commits. Every commit is noted with a
'-'

- [ganetimgr1.7] Fix problematic migrations

The migrations were messed up, possibly because since the Django 1.7 branch
was introduced more migrations have been added.

The problem was quite cryptic, kept getting the following error message:
django.core.exceptions.ImproperlyConfigured:
	Application labels aren't unique, duplicates: apply

Searches kept pointing to ways to actually rename your app if it has the same
label with another one, as Django 1.7 no longer allows this. This was not useful
though, as no duplicate app existed. Not really sure what the problem actually was,
but deleting existing migrations in the app fixed it.

Afterwards, though, we could not create migrations for certain apps (eg. notifications).
This happened as notifications had already some south migrations, probably trash left while
merging the commits from 1.7 branch.

To fix this:

1) All (django) migrations were deleted
2) All south-migrations for every app were added to <app>/south_migrations directory.
3) ./manage.py makemigrations to create new initial django migrations
4) ./manage.py migrate to (fake) apply the new migrations.

- [ganetimgr1.8] Add missing `auditlog` initial migrations

Interesting note (from docs):

https://docs.djangoproject.com/en/dev/topics/migrations/#adding-migrations-to-apps

    Adding migrations to new apps is straightforward - they come preconfigured to accept
    migrations, and so just run makemigrations once you’ve made some changes.

    If your app already has models and database tables, and doesn’t have migrations yet
    (for example, you created it against a previous Django version), you’ll need to
    convert it to use migrations; this is a simple process:

    $ python manage.py makemigrations your_app_label

Adding missing `auditlog` migrations

- [ganetimgr1.8] Remove `south_migrations`

No longer needed

- [ganetimgr1.8] Changes in models, additional migrations

Refactor the action function in 'apply' admin model

This patch refactors the action method of the 'apply/admin.py' module in
order to improve the code's readability without affecting the core
functionality.

In details, we subtract the action helper method '_submit_applications'
from the 'make_fast_create_actions' method in a separate method named
'make_submit_applications'. This method new is called using the
'functools.partial()' method. Finally, we also renamed the
'ApplicationAdmin' model to 'InstanceApplicationAdmin' since it refers
to the InstanceApplication model.

Conflicts:
	apply/admin.py

- [ganetimgr1.8] Remove `null=True` from `ManyToManyField`

We kept getting the following:

* apply.Organization.users: (fields.W340) null has no effect on ManyToManyField
* ganeti.Network.groups: (fields.W340) null has no effect on ManyToManyField

Removed `null=True` as Django 1.8 now suggests.
This produced migrations too, in apps `ganeti` and `apply`. Especially `ganeti` app
had no initial migrations so they were created too. If this operation (changing this
field) produced some actual SQL code we would have to create the initial migrations
and then a second for this operation, but this is not the case since removing `null
= True` does nothing in SQL level.

Interesting issue:

When trying to test migrations for app `ganeti`, e.g. added a new field in a model
and ran `./manage.py makemigrations` we got an error. The issue was that in
`apply/admin.py` we referenced models from `ganeti` app in module-import time.

This is a problem in Django > 1.7.

Fix: Removed the relevant code (the one populating the admin ListView actions)
in `apply/admin.py` since it was no longer relevant. Issue fixed.

Messages of the old commits follow: