Refactor the action function in 'apply' admin model
This patch refactors the action method of the 'apply/admin.py' module in
order to improve the code's readability without affecting the core
functionality.

In details, we subtract the action helper method '_submit_applications'
from the 'make_fast_create_actions' method in a separate method named
'make_submit_applications'. This method new is called using the
'functools.partial()' method. Finally, we also renamed the
'ApplicationAdmin' model to 'InstanceApplicationAdmin' since it refers
to the InstanceApplication model.

Conflicts:
	apply/admin.py