[ganetimgr1.8] Standard 1.4 - 1.8 changes
Changes summary:

* Change imports for `urls`, `patterns`, `include`
    (`django.conf.urls.defaults` -> `django.conf.urls`)
* Change exception message retrieval
    (`Exception.message[0]` -> `Exception.args[0][0]`)
* Change `url` templatetag, surround names with quotes
* Change `request.user.get_profile()` to `request.user.userprofile`
* Change `sha` to `hashlib`, use `hashlib.sha1` instead of `sha`
* Add `SESSION_SERIALIZER` to `settings.py.dist`
* Change `mimetype` kwarg to `content_type` in `HttpResponse`
* Change `setup_environ` to `os.environ.set_default`
* Use Python's `json` instead of `django.utils.simplejson` (deprecated)
* Replace `gevent.pool.imap` with its map equivalent (see commit below)
* Replace `django.db.close_connection` with `django.db.close_old_connections`
* Replace `forms.utils.ValidationError` with `django.core.exceptions.ValidationError`
* Add `ALLOWED_HOSTS` in `settings.py.dist`

Messages below provide more details

[ganetimgr1.7] Start merging & resolving conflicts

Starting the upgrade to Django 1.7. As a lot of time has passed since
this was implemented, we expect to have various conflicts.

upgrade to django 1.5

Conflicts:
	accounts/urls.py
	regbackends/ganetimgr/__init__.py
	templates/registration/registration_complete.html

[ganetimgr1.7] commit merge

upgrade to django 1.6

Conflicts:
	ganeti/models.py

change mimetype keyword to content_type in HttpResponse kwargs

Remove the deprecated 'setup_environ' method call

This patch removes the call to setup_environ() method from the watcher
module, that has been removed from the django.core.management module in
Django v1.6.

Use json module everywhere in the project

Django v1.5 deprecates 'django.utils.simplejson' in favor of the
built-in python's json module.

This patch updates the project's code to use the json module everywhere.

Conflicts:
	django_markdown/templatetags/django_markdown.py
	django_markdown/utils.py
	docs/source/install.rst
	ganeti/views.py

Replace gevent.pool.imap with its map equivalent

In gevent v1.0, the implementation of the 'gevent.pool.imap' has been
modified to return an iterator as its result. In ganetimgr, this
modification results in no data shown to the user since this iterator
is never consumed by the application.

To fix this issue, we replace the call to imap with the map equivalent,
which makes an internal call the to imap method and then consumes it
too. Finally, we also removed the call to 'gevent.pool.join' method
since it is not necessary anymore.

Conflicts:
	ganeti/models.py
	ganeti/views.py
	notifications/views.py

[ganetimgr1.7] Replace deprecated close_connection() with close_old_connections()

[ganetimgr1.7] Remove deprecated `ValidationError` imports

`ValidationError` used to be imported from `forms.utils` or
`forms.util`. This is deprecated. Import it from `django.core.exceptions`.

Add the ALLOWED_HOST option in settings.py.dist

Since Django v1.5, the ALLOWED_HOSTS option is required in production.

The new ALLOWED_HOSTS setting validates the request's Host header and
protects against host-poisoning attacks. This setting is now required
whenever DEBUG is set to False. For more details see:

https://docs.djangoproject.com/en/1.8/ref/settings/#allowed-hosts

Conflicts:
	ganetimgr/settings.py.dist