[vmsecnode] Display instance's secondary node(s)
Instances may be located in multiple nodes. This gets the related
information for an instance's secondary nodes from the GanetiRAPI
and adds a related field in the frontend to display any additional
nodes the instance is located on (if they exist).

This information is visible only on admins and users that have the
`ganeti.view_instances` permission.