[showapproved] Show administrator who approved an InstanceApplication
The administrator who approved an ``InstanceApplication`` object is not currently
stored/visible anywhere. This adds a field in the ``InstanceApplication`` model,
``reviewer`` and implements the required logic.

Reviewer is displayed both in frontend (in application details) and is visible
in the mail sent to the site admins after successful instance creation.

Ref: T428