Support extstorage providers
First Step in supporting Instance creation with ExtStorage providers
enabled in clusters. Extstorage is the 'ext' disk template but needs to
specify the name of the provider in the disks dictionary.

(A cluster can have multiple ExtStorage providers enabled)

Ganeti has gnt-storage as a CLI interface for extstorage that closely
resmbleds the gnt-os interface but it (gnt-storage) is not exposed
through the API.

Taking a similar approach with alternate VGs for DRBD template, we
designate a cluster-wide tag GANETI_TAG_PREFIX:ext:<provider_name> as a
way for Ganetimgr to determine the name of all available Extstorage
providers.