Fix display of "Applications" in sidebar
The code only permitted users with superuser privileges to see
the page listing `InstanceApplication`s. Seemed like a bug though,
since inner `if` statement seemed to be designed to allow access
to users with `can_handle_applications` or `can_view_applications`
too, but since outer `if` required the user to be a superuser the
inner condition was never evaluated.o

NOTE: `can_handle_applications`, `can_view_applications` are values
manually injected in the `request` object from a context processor.

Ref: T847